qa_automation/requirements.txt
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
decorator==5.1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
geopy==2.4.1
h11==0.14.0
idna==3.4
jsonpath-ng==1.5.3
jsonpath-rw==1.4.0
jsonpath-rw-ext==1.2.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
natsort==8.4.0
opencv-python==4.10.0.84
openpyxl==3.1.2
outcome==1.2.0
ply==3.11
psycopg2==2.9.7
pycparser==2.21
pyjsonselect==0.2.2
pyparsing==3.1.2
PySocks==1.7.1
pytz==2024.2
referencing==0.30.0
requests==2.31.0
robotframework==6.1
robotframework-databaselibrary==1.3.1
robotframework-excellib==2.0.1
robotframework-jsonlibrary==0.5
robotframework-pabot==2.16.0
robotframework-pythonlibcore==4.4.1
robotframework-requests==0.9.5
robotframework-seleniumlibrary==6.5.0
robotframework-stacktrace==0.4.1
rpds-py==0.9.2
selenium==4.25.0
six==1.16.0
shapely==2.0.6
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.12.2
urllib3==1.26.16
websocket-client==1.8.0
wsproto==1.2.0